requirements.txt
azure-functions
azure-identity
azure-mgmt-resourcegraph
azure-storage-blob